requirements.txt
python-dotenv==1.0.1
loguru==0.7.3
mcrcon==0.7.0
redis==5.2.1
psycopg2==2.9.10